[~][refacto][2048][random-cell-after-actions] update files tree, add actionsTypes & arrowUp folders move moveUp function in up.js file, switch function syntax to ES6
