+++ src/functions/actionsTypes/arrowUp/up.js
@@ -1,12 +1,13 @@
+import { cellsArray } from '../../../index.js';
+import addRandomValue from '../../addRandomValue.js';
 /**
- *
  * @param {int} row
  * @param {int} col
  * @param {Node} node
  * @param {string} type
  * @returns  recursive function
  */
- const upCell = (row, col, node) =>{
+ const moveUp = (row, col, node) =>{
 	if(row === 1) return;
 
 	let newPos = row - 1 + '-' + col;
@@ -28,7 +29,21 @@
 		}
 	}
 	next.appendChild(node);
-	return upCell(row - 1, col, node);
+	return moveUp(row - 1, col, node);
 }
 
-export default upCell
+/**
+ * @param {Array} active
+ */
+const up = (active) => {
+	for (const cell of active) {
+		let row = cell.parentNode.id.split('-')[0];
+		let col = cell.parentNode.id.split('-')[1];
+		if (row > 1) {
+			moveUp(row, col, cell);
+		}
+	}
+  addRandomValue(cellsArray, 1)
+}
+
+export default up
